Switch to Maven + Metrics + SimpleSpec
