--- octobot
@@ -1,17 +1,0 @@
-#!/bin/bash
-
-options[0]="configFile=`pwd`/example-config.yml"
-options[1]="log4j.configuration=`pwd`/log4j.properties"
-
-JVM_OPTS="-server"
-CLASSPATH="-cp ../lib/*:.:../tasks.jar"
-EXECUTABLE="com.urbanairship.octobot.Octobot"
-
-for i in "${options[@]}"
-do
-  JVM_OPTS="$JVM_OPTS -D$i "
-done
-
-echo -e "Launching Octobot...\n"
-echo -e "Options specified: $JVM_OPTS\n"
-cd build && java $JVM_OPTS $CLASSPATH $EXECUTABLE
