+++ src/test/scala/com/urbanairship/octobot/TestUtils.scala
@@ -1,0 +1,9 @@
+import java.io._
+import java.net._
+
+import org.apache.log4j.Logger;
+import org.apache.log4j.BasicConfigurator;
+
+object TestUtils {
+
+}
