--- test/mocks/build_mocks
@@ -1,8 +1,0 @@
-echo "Compiling mock tasks..."
-cd ../..
-javac -cp jar/octobot.jar test/mocks/com/urbanairship/octobot/tasks/*.java
-
-echo "Bottling mock tasks in mocks.jar..."
-cd test/mocks
-jar -cvf mocks.jar com/urbanairship/octobot/tasks/*.class
-echo "Done"
