--- test/specs/TestUtils.scala
@@ -1,12 +1,0 @@
-import org.scalatest.Spec
-import org.scalatest.matchers.MustMatchers
-
-import java.io._
-import java.net._
-
-import org.apache.log4j.Logger;
-import org.apache.log4j.BasicConfigurator;
-
-object TestUtils {
-
-}
